Fixing Reset Button Bug: Restoring Field Interaction
The change ensures that after pressing the reset button, players can once again click on the game board to start a new game. Previously, there was a bug preventing players from clicking on the fields after resetting the game, which was introduced by the last update "Preventing Game Result Change".